refactor(web): fold provider runtime into core
