+++ .oxlintrc.json
@@ -243,7 +243,6 @@
         "packages/terminal-core/**/*.ts",
         "packages/normalization-core/**/*.ts",
         "packages/model-catalog-core/**/*.ts",
-        "packages/web-content-core/**/*.ts",
         "packages/agent-core/**/*.ts",
         "packages/acp-core/**/*.ts",
         "packages/ai/**/*.ts",
